taken from https://github.com/luciotato/waitfor-ES6/blob/master/waitfor-demo.js
Dummy blog server. Exposes "mode of employ"
The "mode of employ" of wait.for is to launch a fiber when you need to launch a long process to attend a request. Ideally here:
var server = http.createServer(
  function(req, res){
    console.log('req!');
    wait.launchFiber(handler,req,res); //handle in a fiber
    // keep node spinning
  }).listen(8000);
  
after that,inside function handler(req,res) and every function you cal from there, you'll be able to use wait.for(ayncFn...